Move options into core
