+++ tests/flytekit/unit/core/test_launch_plan.py
@@ -7,13 +7,14 @@ from flyteidl.admin import launch_plan_pb2 as _launch_plan_idl
 import flytekit.configuration
 from flytekit.configuration import Image, ImageConfig
 from flytekit.core import launch_plan, notification
+from flytekit.core.options import Options
 from flytekit.core.schedule import CronSchedule
 from flytekit.core.task import task
 from flytekit.core.workflow import workflow
 from flytekit.models.common import Annotations, AuthRole, Labels, RawOutputDataConfig
 from flytekit.models.core import execution as _execution_model
 from flytekit.models.core import identifier as identifier_models
-from flytekit.tools.translator import get_serializable, Options
+from flytekit.tools.translator import get_serializable
 
 default_img = Image(name="default", fqn="test", tag="tag")
 serialization_settings = flytekit.configuration.SerializationSettings(
